Change CORS response headers and cookie setting
Access-Control-Allow-Credentials header is now set on all responses when
the CORS configuration is set (see bed50db). It is set to `true`, which
means cookies will be sent on cross-origin requests.

The `jwt` cookie for the authentication token is now set with a domain
value from the NIGHTSHADES_COOKIE_DOMAIN environment variable. This
should be set to something like `.nightshades.xyz` so that the client
can access these cookies despite not living on api.nightshades.xyz