Add an alternate response for /auth completion
This adds support for opening the /auth endpoint in a browser dialog
window.

If a ?postMessage=true parameter is given on the start of the /auth
flow, and intermediate cookie will be set. This cookie will inform the
completion stage of the flow to return a JS response instead of a JSON
one. This JS response will call postMessage on the opening window to
note that authentication has completed.

A NIGHTSHADES_PUBLIC_ORIGIN environment variable is required to specify
the target origin of the postMessage call (see targetOrigin at
https://developer.mozilla.org/en-US/docs/Web/API/Window/postMessage).