Change api.start_unit to return dict of unit
Previously, this only returned the unit ID. A better way to do this
change would be to use the `.returning()` clause of the `Unit.insert`
call—which should return all columns of the model, according to the
peewee documentation. However, I only seem to be able to get this to
return the ID. I could do a hacky string replace on the SQL, but instead
I'm just going to do a SELECT query after the INSERT query.